requirements.txt
matplotlib==3.8.2
numpy==1.26.3
numpy_financial==1.0.0
pandas==2.1.4
pydantic==2.5.3
pytest==7.4.4
python-dotenv==1.0.0
python_dateutil==2.8.2
Requests==2.31.0
streamlit==1.30.0
